Started with Layer Stack
Implemented (very) basic layer and implementation. Added (still untested) pooling and movments of commands lists for one frame. Still missing the ability to dispatch command lists witch are not needed in the next frame

+++ src/UGF12/DirectX/XCmdList.h
@@ -97,6 +97,6 @@ namespace GxDirect {
 			/// <summary>
 			/// Value to wait for on execution
 			/// </summary>
-			UINT m_uiExecutionValue = 0;
+			UINT64 m_uiExecutionValue = 0;
 	};
 }
